requirements.txt
anyio==3.7.0
bert-extractive-summarizer==0.10.1
blis==0.7.9
boto3==1.26.144
botocore==1.29.144
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
dnspython==2.3.0
email-validator==2.0.0.post2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
exceptiongroup==1.1.1
fastapi==0.95.2
filelock==3.12.0
flashtext==2.7
fsspec==2023.5.0
future==0.18.3
gensim==4.3.1
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.14.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.24.3
orjson==3.8.14
packaging==23.1
pandas==2.0.2
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
pydantic==1.10.8
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3
pywsd==1.2.5
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
starlette==0.27.0
sympy==1.12
textwrap3==0.9.2
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
typer==0.7.0
typing_extensions==4.6.2
tzdata==2023.3
ujson==5.7.0
Unidecode==1.3.6
urllib3==1.26.16
uvicorn==0.22.0
wasabi==1.1.1
watchfiles==0.19.0
websockets==11.0.3
wn==0.0.23
